move add/sub macros to macros module
